requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
defusedxml==0.7.1
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
filelock==3.17.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==0.3.4
langchain-core==0.3.31
langchain-openai==0.2.9
langgraph==0.2.53
langgraph-checkpoint==2.0.18
langgraph-sdk==0.1.55
langsmith==0.1.146
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
numpy==1.26.4
openai==1.65.5
orjson==3.10.15
packaging==24.2
pillow==11.1.0
propcache==0.3.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.9.0
together==1.4.1
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
zstandard==0.23.0

#cairosvg